Add schema management to groups panel and remove standalone SchemaView
