+++ gerenciador_postgres/gui/main_window.py
@@ -603,7 +603,7 @@ class MainWindow(QMainWindow):
         if not self.groups_controller:
             raise RuntimeError('Não conectado')
         from .groups_view import PrivilegesView
-        return PrivilegesView(controller=self.groups_controller)
+        return PrivilegesView(controller=self.groups_controller, schema_controller=self.schema_controller)
 
     def _factory_schema_privileges(self):
         if not self.schema_controller or not self.groups_controller:
